Machine Learning/Intelligent doc and image scanner/requirements.txt
opencv-python>=4.7.0.72
pytesseract>=0.3.10
easyocr>=1.6.2
pandas>=2.2.0
numpy>=1.26.0
Pillow>=9.5.0
pdf2image>=1.16.3
tqdm>=4.65.0
textblob>=0.17.1
rapidfuzz>=3.0.0
spacy>=3.6.0
streamlit>=1.28.0
transformers>=4.40.0
torch
torchvision
torchaudio
